feat: Improve code organization (types-refactorings; `@/` import schema)
