+++ src/data/PSP22_ASSETS.json
