+++ src/helpers/getAbi.ts
@@ -1,2 +1,2 @@
-import PSP22_Abi from '../psp22_abi.json'
+import PSP22_Abi from '../metadata/psp22.json'
 export const psp22Abi = PSP22_Abi as Record<string, any>
