+++ src/types/ChainAsset.ts
@@ -1,8 +1,3 @@
-export enum AssetType {
-  PSP22 = 'PSP22',
-  PSP34 = 'PSP34',
-}
-
 export interface ChainAsset {
   originChain: string
   slug: string
@@ -12,4 +7,9 @@ export interface ChainAsset {
   assetType: AssetType
   metadata: Record<any, any> | null
   iconPath: string
+}
+
+export enum AssetType {
+  PSP22 = 'PSP22',
+  PSP34 = 'PSP34',
 }
