+++ src/types/index.ts
@@ -1,0 +1,6 @@
+// @index('./*.ts', f => `export * from '${f.path}'`)
+export * from './ChainAsset'
+export * from './SubstrateDeployment'
+export * from './SubstrateWallet'
+export * from './UseInkathonProviderContext'
+// @endindex
